requirements.txt
graphene>=2.1,<3
graphene-django>=2.1,<3
graphql-core>=2.1,<3
django==3.1.8
django-filter>=2
psycopg2==2.8.6
gevent==21.1.2
